# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ketringan v2 API Documentation",
        "description": "API documentation for Ketringan App v2. Usable for PWA and Android Mobile",
        "contact": {
            "email": "ilhamfi_2701@yahoo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "localhost:8000",
            "description": "Main development server"
        }
    ],
    "paths": {
        "/api/v1/banner/active_banner": {
            "get": {
                "description": "Get All Active Banner",
                "operationId": "App\\Http\\Controllers\\api\\v1\\BannerController::active_banner",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Get Banner with active status true or false, default status is true",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    }
                }
            }
        },
        "/api/v1/konsumen/request_membership": {
            "post": {
                "description": "Request membership for konsumen",
                "operationId": "App\\Http\\Controllers\\api\\v1\\KonsumenController::membership_request",
                "parameters": [
                    {
                        "name": "No_Telfon",
                        "in": "query",
                        "description": "User's phone number",
                        "required": true
                    },
                    {
                        "name": "Alamat",
                        "in": "query",
                        "description": "User's address",
                        "required": true
                    },
                    {
                        "name": "Catatan",
                        "in": "query",
                        "description": "User's note",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    },
                    "401": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/konsumen/activated_membership": {
            "get": {
                "description": "Get activated membership for konsumen",
                "operationId": "App\\Http\\Controllers\\api\\v1\\KonsumenController::get_activated_membership",
                "responses": {
                    "200": {
                        "description": "Request OK"
                    }
                }
            }
        },
        "/api/v1/menu": {
            "get": {
                "description": "Get All Menu",
                "operationId": "App\\Http\\Controllers\\api\\v1\\MenuController::all",
                "parameters": [
                    {
                        "name": "id_region",
                        "in": "query",
                        "description": "Get Menu Where Id Region is x",
                        "required": false
                    },
                    {
                        "name": "id_kategori",
                        "in": "query",
                        "description": "Get Menu Where Id Kategori is x",
                        "required": false
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Get Menu Where Max Price is x",
                        "required": false
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Get Menu Where Min Price is x",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    }
                }
            }
        },
        "/api/v1/testimoni/active_testimoni": {
            "get": {
                "description": "Get All Active Testimoni",
                "operationId": "App\\Http\\Controllers\\api\\v1\\TestimoniController::active_testimoni",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Get Testimoni with active status true or false, default status is true",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "description": "API for Login, the login will be generate token that must used by all pages that require authentication",
                "operationId": "App\\Http\\Controllers\\api\\v1\\UserController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "description": "API for Login, the login will be generate token that must used by all pages that require authentication",
                "operationId": "App\\Http\\Controllers\\api\\v1\\UserController::register",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "User's name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true
                    },
                    {
                        "name": "no_telefon",
                        "in": "query",
                        "description": "User's phone number",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Chosen password from user",
                        "required": true
                    },
                    {
                        "name": "password_confirm",
                        "in": "query",
                        "description": "Password confirmation",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    },
                    "401": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/v1/token_confirmation/{id}": {
            "get": {
                "description": "Token confirmation after registering",
                "operationId": "App\\Http\\Controllers\\api\\v1\\UserController::token_confirmation",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Users registration Token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    },
                    "401": {
                        "description": "Invalid verification token"
                    }
                }
            }
        },
        "/api/v1/details": {
            "get": {
                "description": "Get user detail",
                "operationId": "App\\Http\\Controllers\\api\\v1\\UserController::details",
                "responses": {
                    "200": {
                        "description": "Request OK"
                    },
                    "401": {
                        "description": "Invalid verification token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/v1/vendor/partnership_request": {
            "post": {
                "description": "Request partnership for vendor",
                "parameters": [
                    {
                        "name": "No_Telfon",
                        "in": "query",
                        "description": "User's phone number",
                        "required": true
                    },
                    {
                        "name": "Alamat",
                        "in": "query",
                        "description": "User's address",
                        "required": true
                    },
                    {
                        "name": "Catatan",
                        "in": "query",
                        "description": "User's note",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request OK"
                    },
                    "401": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}